Complete user login page and created domUpdates file
