+++ src/scripts.js
@@ -1,23 +1,20 @@
 import './css/base.scss';
-// An example of how you tell webpack to use an image (also need to link to it in the index.html)
-// import './images/turing-logo.png'
+import './images/IMG_3938.jpg'
 import Customers from './classes/Customers.js';
 import Bookings from './classes/Bookings.js';
 import Rooms from './classes/Rooms.js';
 import {customersAPI, roomsAPI, bookingsAPI, updateBookingsAPI} from './apiCalls';
+import {hide, show, showViewHome, showViewNewBooking, showViewSuccessfullyBookedRoom, showBookingDetails, showResult, showCustomerDashboard, showLoginErrorMessage, clearUserBookingResults, displayTotalSpent, clearAvailableRoomsToBook, showListingsMessageAndAvailableRoomsToBookView, showErrorListingsMessage, showListingsMessageAndAvailableRoomsToBook, showRoom, uppercaseFirstLetter} from './domUpdates';
 
 let customers;
 let rooms;
 let bookings;
-let user;
 
 Promise.all([customersAPI, roomsAPI, bookingsAPI]).then(data => {
   customers = data[0].customers.map(customer => new Customers(customer))
-  // console.log('customers: ', customers)
-  user = customers[getRandomIndex(customers)];
-  // console.log("user: ", user)
+  //   console.log('customers: ', customers)
   rooms = data[1].rooms.map(room => new Rooms(room))
-//   console.log('rooms :', rooms)
+  //   console.log('rooms :', rooms)
   bookings = data[2].bookings.map(booking => new Bookings(booking))
   // console.log('bookings: ', bookings)
 }).catch(error => console.log(error));
@@ -30,64 +27,33 @@ const viewHome = document.getElementById('viewHome');
 const viewCustomerDashboard = document.getElementById('viewCustomerDashboard');
 const viewNewBooking = document.getElementById('viewNewBooking');
 const viewSuccessfullyBookedRoom = document.getElementById('viewSuccessfullyBookedRoom');
-let calendar = document.getElementById('calendar');
+const calendar = document.getElementById('calendar');
 const resultsFilter = document.getElementById('resultsFilter');
 const listingsHeaderMessage = document.getElementById('listingsHeaderMessage');
 const availableRoomsToBook = document.getElementById('availableRoomsToBook');
 const bookedRoomOverview = document.getElementById('bookedRoomOverview');
-let currentDate = new Date().toJSON().slice(0,10);
+const loginErrorMessage = document.getElementById('loginErrorMessage');
+const errorSpacer = document.getElementById('errorSpacer');
+let user;
+let currentDate = new Date().toJSON().slice(0, 10);
 document.getElementById('calendar').setAttribute('value', currentDate);
 document.getElementById('calendar').setAttribute('min', currentDate)
 
-const getRandomIndex = (arr) => {
-  return Math.floor(Math.random() * arr.length);
-};
-
-const showViewHome = () => {
-  hide([viewCustomerDashboard, viewNewBooking, resultsFilter, viewSuccessfullyBookedRoom, buttonHome, buttonCurrentBookings, buttonNewBooking]);
-  show([viewHome]);
-};
-//buttonHome, buttonCurrentBookings, buttonNewBooking
-const showViewCustomerDashboard = () => {
-  hide([viewHome, viewNewBooking, resultsFilter, viewSuccessfullyBookedRoom, buttonCurrentBookings]);
-  show([viewCustomerDashboard, buttonHome, buttonNewBooking]);
-  showBookings();
-};
-
-const showViewNewBooking = () => {
-  hide([listingsHeaderMessage, availableRoomsToBook, viewCustomerDashboard, viewHome, resultsFilter, viewSuccessfullyBookedRoom, buttonNewBooking]);
-  show([viewNewBooking, buttonHome, buttonCurrentBookings, ]);
-};
-
-const showViewSuccessfullyBookedRoom = () => {
-  hide([viewNewBooking, viewCustomerDashboard, viewHome, resultsFilter, buttonHome, buttonCurrentBookings, buttonNewBooking]);
-  show([viewSuccessfullyBookedRoom, buttonHome, buttonCurrentBookings, buttonNewBooking]);
-};
-
-const hide = (array) => {
-  array.forEach(element => element.classList.add('hidden'));
-};
-  
-const show = (array) => {
-  array.forEach(element => element.classList.remove('hidden'));
-};
 
 const loginUser = () => {
-  console.log("user is logged in");
-  //form may refresh page?
-};
-
-const showBookingDetails = (pickedRoomNumber, pickedDate) => {
-  return `<h1>Congrats on booking your stay!</h1>
-    <h2>Room Details:</h2>
-    <section class="booked-room-details">
-        Room Number: ${pickedRoomNumber}<br/>
-        Booked Date: ${pickedDate}
-    </section>
-    <div class="bookingDetailsNavigation">
-        <button class="nav-button" id="buttonSeeCurrentBookings">See All Current Bookings</button>
-        <button class="nav-button" id="buttonCreateAnotherNewBooking">Create Another Booking</button>
-    </div>`
+  let usernameInput = document.getElementById('username').value;
+  let passwordInput = document.getElementById('password').value;
+  // console.log(usernameInput)
+  // console.log(passwordInput)
+  let customerPhrase = usernameInput.substring(0, 8);
+  let customerNumber = parseInt(usernameInput.substring(8, 10))
+  if (customerPhrase === 'customer' && customerNumber > 0 && customerNumber < 51 && passwordInput === 'overlook2021') {
+    user = customers[customerNumber - 1]
+    // console.log("user: ", user)
+    showViewCustomerDashboard();
+  } else {
+    showLoginErrorMessage();
+  }
 };
 
 const bookRoom = (event) => {
@@ -97,7 +63,6 @@ const bookRoom = (event) => {
   updateBookingsAPI(userId, pickedDate, pickedRoomNumber).then(data => {
     bookings.push(new Bookings(data.newBooking))
     showViewSuccessfullyBookedRoom();
-    bookedRoomOverview.innerHTML = ``
     bookedRoomOverview.innerHTML = showBookingDetails(pickedRoomNumber, pickedDate)
     const buttonSeeCurrentBookings = document.getElementById('buttonSeeCurrentBookings');
     const buttonCreateAnotherNewBooking = document.getElementById('buttonCreateAnotherNewBooking');
@@ -106,27 +71,9 @@ const bookRoom = (event) => {
   });
 };
 
-const showResult = (item) => {
-  return `<section class="currently-booked-room"> Date: ${item.date}<br/>
-    Room Number: ${item.roomNumber}</section>`;
-};
-
-const uppercaseFirstLetter = (item) => {
-  let phrase = item.split(" ").map((letter) => letter.charAt(0).toUpperCase() + letter.substring(1)).join(' ');
-  return phrase;
-};
-
-const showRoom = (item, date) => {
-  return `
-    <section class="available-rooms-to-book">
-    Room Number: ${item.number}<br/>
-    Type: ${uppercaseFirstLetter(item.roomType)}<br/>
-    Bed Size: ${uppercaseFirstLetter(item.bedSize)}<br/>
-    Number of Beds: ${item.numBeds}<br/>
-    Cost Per Night: $${item.costPerNight.toFixed(2)}<br/>
-    Bidet: ${item.bidet ? "Yes" : "No"}<br/>
-    <button data-user-id="${user.id}" data-picked-date="${date}" data-picked-room-number="${item.number}" class="nav-button" buttonBookStay>Book Room</button>
-    </section>`
+const showViewCustomerDashboard = () => {
+  showCustomerDashboard();
+  showBookings();
 };
 
 const newBookingsAPI = () => {
@@ -138,10 +85,8 @@ const showBookings = () => {
     bookings = response.bookings.map(booking => new Bookings(booking))
     const userBookings = document.getElementById('userBookings');
     const totalSpent = document.getElementById('totalSpent');
-    totalSpent.innerHTML = '';
-    userBookings.innerHTML = '';
+    clearUserBookingResults();
     let result = bookings.filter(booking => booking.userID === user.id);
-    console.log(result);
     result.forEach(item => {
       userBookings.innerHTML += showResult(item);
     })
@@ -150,7 +95,7 @@ const showBookings = () => {
       acc += matchingRoom.costPerNight;
       return acc;
     }, 0)
-    totalSpent.innerHTML = `Total Spent: $${totalCost.toFixed(2)}`
+    displayTotalSpent(totalCost);
   }).catch(error => console.log(error));
 }
 
@@ -159,7 +104,7 @@ const selectDate = () => {
   let date = pickedUserDate.replace('-', '/').replace('-', '/');
   let resultsFilter = document.getElementById('resultsFilter');
   let filterSelector = resultsFilter.querySelector('input:checked');
-  availableRoomsToBook.innerHTML = '';
+  clearAvailableRoomsToBook();
   let result = rooms.filter(room => {
     let takenRooms = bookings.filter(booking => booking.roomNumber === room.number);
     return !takenRooms.find(room => {
@@ -168,20 +113,15 @@ const selectDate = () => {
   });
   if (filterSelector.value) {
     result = result.filter(room => {
-      console.log(room.roomType)
-      console.log(filterSelector.value)
       return room.roomType === filterSelector.value
     });
   }
-  show([listingsHeaderMessage, availableRoomsToBook])
+  showListingsMessageAndAvailableRoomsToBookView();
   if (result.length === 0) {
-    listingsHeaderMessage.innerHTML = `I'm sorry, but there are no rooms available for that date. Please search another date.`
-    availableRoomsToBook.innerHTML = '';
+    showErrorListingsMessage();
   } else {
     result.forEach(room => {
-      show([resultsFilter]);
-      listingsHeaderMessage.innerHTML = `Available Rooms To Book For The Selected Day:<br/>`
-      availableRoomsToBook.innerHTML += showRoom(room, date);
+      showListingsMessageAndAvailableRoomsToBook(room, date, user);
     });
   }
 };
